wip: add gauges context
