+++ src/features/Gauges/weightDistribution.js
@@ -1,0 +1,53 @@
+import React, { memo } from "react";
+import { Card } from "react-bootstrap";
+import { Doughnut } from "react-chartjs-2";
+import styled from "styled-components";
+
+const StyledCard = styled(Card)`
+  padding: 10px;
+  margin: 5px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
+
+const data = {
+  labels: ["Red", "Blue", "Yellow", "Green", "Purple", "Orange"],
+  datasets: [
+    {
+      label: "# of Votes",
+      data: [12, 19, 3, 5, 2, 3],
+      backgroundColor: [
+        "rgba(255, 99, 132, 0.2)",
+        "rgba(54, 162, 235, 0.2)",
+        "rgba(255, 206, 86, 0.2)",
+        "rgba(75, 192, 192, 0.2)",
+        "rgba(153, 102, 255, 0.2)",
+        "rgba(255, 159, 64, 0.2)"
+      ],
+      borderColor: [
+        "rgba(255, 99, 132, 1)",
+        "rgba(54, 162, 235, 1)",
+        "rgba(255, 206, 86, 1)",
+        "rgba(75, 192, 192, 1)",
+        "rgba(153, 102, 255, 1)",
+        "rgba(255, 159, 64, 1)"
+      ],
+      borderWidth: 1
+    }
+  ]
+};
+
+const WeightDistribution = () => {
+  return (
+    <StyledCard style={{ width: 277 }}>
+      <Card.Title>Weight distribution</Card.Title>
+      <Card.Body className="w-100">
+        <Doughnut data={data} style={{ width: 277 }} />
+      </Card.Body>
+    </StyledCard>
+  );
+};
+
+export default memo(WeightDistribution);
